backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
supabase==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
requests==2.31.0
